Move `isMSAADebugLoggingEnabled` from IAccessibleHandler to IAccessibleHandler.utils to workaround a circular import
